+++ no-morph-test.ts
@@ -1,0 +1,8 @@
+interface UserData {
+  id: string;
+  name: string;
+}
+
+function processUser(user: UserData) {
+  return user.name;
+}